restructure app, and test replace, and native addons
