+++ test/index.sh
@@ -4,8 +4,7 @@
 # use noderify to bundle itself, then check that the bundled
 # version can still bundle noderify and get the exact same bundle.
 
-okay () {
-  "$@"
+was_okay () {
   if [ $? -eq 0 ]; then
     echo Okay
   else
@@ -13,6 +12,12 @@ okay () {
     exit 1
   fi
 }
+
+okay () {
+  "$@"
+  was_okay
+}
+
 
 not_okay () {
   "$@"
@@ -36,14 +41,30 @@ echo $PWD
 tidy
 
 noderify=./index.js
-for file in test/*.js; do
+#//
+#//  echo $noderify  "$file"
+#//  time $noderify "$file" > _bundle.js
+#//
+#//  okay node _bundle.js
 
-  echo $noderify  "$file"
-  time $noderify "$file" > _bundle.js
+echo "Test module not found error:"
+not_okay $noderify test/missing.js > /dev/null
 
-  okay node _bundle.js
+$noderify test/self-reference.js > _bundle.js 
+was_okay
+okay node _bundle.js
 
-done
+echo "TEST NOT FOUND, but replace missing module"
+$noderify test/missing.js --replace.foo-bar-baz=./fbb > _bundle.js 
+was_okay
+okay node _bundle.js
+
+$noderify test/native.js --filter sodium-native --filter leveldown > _bundle.js 
+was_okay
+okay node _bundle.js
+
+
+#done
 
 #set -e # exit with an error if any command fails
 
@@ -59,8 +80,10 @@ shasum _b.js _b2.js
 okay diff _b.js _b2.js
 
 not_okay node _b.js missing.js  2> /dev/null > /dev/null
+echo "ignore missing"
 okay node _b.js missing.js --ignore-missing 2> /dev/null > /dev/null
 
 tidy
+
