+++ test/native.js
@@ -1,0 +1,3 @@
+
+require('sodium-native')
+require('leveldown')
